Create FishSimulationGame
A lab done that reads in a file indicating the number of each object (fish and food) to create and the starting positions of all objects. There is a 1% chance that a new food will be added to the screen. If the fish eats the food, it will be deleted from the screen.